Converted test_actions to a python module.
